move the brain region tree from v2 to brain-region-hierarchy
